refactor: Refactor common hooks and mock data structure
