+++ src/entities/store/model/index.ts
@@ -1,1 +1,2 @@
+export * from "./constant";
 export * from "./types";
