+++ src/shared/lib/hooks/useInfiniteScroll.ts
@@ -1,0 +1,45 @@
+"use client";
+
+import { useEffect, useRef } from "react";
+
+interface UseInfiniteScrollOptions {
+  fetchNextPage: () => void;
+  hasNextPage?: boolean;
+  isFetchingNextPage?: boolean;
+  enabled?: boolean;
+}
+
+export function useInfiniteScroll({
+  fetchNextPage,
+  hasNextPage,
+  isFetchingNextPage,
+  enabled = true,
+}: UseInfiniteScrollOptions) {
+  const loadMoreTriggerRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    if (!enabled || !hasNextPage || isFetchingNextPage) return;
+
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting) {
+          fetchNextPage();
+        }
+      },
+      { threshold: 1.0 },
+    );
+
+    const trigger = loadMoreTriggerRef.current;
+    if (trigger) {
+      observer.observe(trigger);
+    }
+
+    return () => {
+      if (trigger) {
+        observer.unobserve(trigger);
+      }
+    };
+  }, [enabled, hasNextPage, isFetchingNextPage, fetchNextPage]);
+
+  return loadMoreTriggerRef;
+}
