+++ src/shared/lib/index.ts
@@ -1,3 +1,4 @@
+export * from "./hooks/useInfiniteScroll";
 export * from "./hooks/useMobile";
 export * from "./utils/authUtils";
 export * from "./utils/formatDateUtils";
